WIP. Adds failing test for multiple date ranges for JSON formatter.
The test report added here  was manually modified, so the mimumims,
totals, etc. are not correct. This was done to shorten the test
assertions and to get around an issue with string to double parsing
in serde_json (https://github.com/serde-rs/json/issues/128) that made
assertions around the many-decimaled values unreliable.