Adds support for formatting reports containing data for multiple date ranges. (#6)
* Does some refactoring to make functions shorter. They are doing some
funky mutable reference passing that we may want to eliminate in the future.
* Switches from Joinery to Itertools because  Itertools was able to handle
joining the result of flat_value_iterator, which is a FlatMap.
* The test report added here was manually modified, so the minimums,
totals, etc. are not correct. This was done to shorten the test
assertions and to get around an issue with string to double parsing
in serde_json (https://github.com/serde-rs/json/issues/128) that made
assertions around the many-decimaled values unreliable.